NP761-9: Add/Remove Admin Roles

+++ server/src/Capisso/Services/UserService.cs
@@ -11,6 +11,7 @@ using System.Threading.Tasks;
 using Capisso.Dto;
 using Capisso.Mapper;
 using System.Text.RegularExpressions;
+using Capisso.Entities;
 
 namespace Capisso.Services
 {
@@ -67,6 +68,11 @@ namespace Capisso.Services
             var user = await _unitOfWork.UserRepository.GetByIdAsync(id) ??
                        throw new EntityNotFoundException($"User with id <{id}> not found");
 
+            if (user.UserRole.Equals(UserRole.Admin) && !(await CheckAdminExistsAfterUserRemoval(user)))
+            {
+                throw new NoAdminExistsException();
+            }
+
             _unitOfWork.UserRepository.Delete(user);
             await _unitOfWork.SaveAsync();
         }
@@ -94,6 +100,27 @@ namespace Capisso.Services
             await _unitOfWork.SaveAsync();
 
             return user.Id;
+        }
+
+        public async Task UpdateUser(UserDto userDto)
+        {
+            var user = UserMapper.FromDto(userDto);
+
+            if (!user.UserRole.Equals(UserRole.Admin) && !(await CheckAdminExistsAfterUserRemoval(user)))
+            {
+                throw new NoAdminExistsException();
+            }
+
+            _unitOfWork.UserRepository.Update(user);
+            await _unitOfWork.SaveAsync();
+        }
+
+        private async Task<bool> CheckAdminExistsAfterUserRemoval(User removedUser)
+        {
+            var remainingAdminList = await _unitOfWork.UserRepository
+                    .FindByAsync(u => u.UserRole.Equals(UserRole.Admin) && !String.Equals(removedUser.Email.ToLower(), u.Email.ToLower()));
+
+            return remainingAdminList.Any();
         }
     }
 }
